Implement filters on deployment, job & origin
This allows filtering of events based on the above mentioned fields.
Multiple filters can be configured and they will run against a message
in the order specified.

A filter has the form

  field₀:comperator₀:value₀;field₁:comperator₁:value₁;…;fieldₙ,comperatorₙ,valueₙ

`field` is the messages field to check against.
`comperator` is how to compate the `field`'s value to the user provided value ('mustContain', 'mustNotContain')
`value` is the value the user want's to check messages for

Benchmarks have been added to get an understanding of the performance
impact of those filters.